fix: gate verde salvo el check del loop
- clang-format con InsertBraces y clang-tidy limpio en los cuatro .cpp.
- Tamaños de std::array convertidos explicitamente a size_t: GCC 12, que es el
  compilador de la imagen de deploy, trata la conversion desde int como
  -Werror=sign-conversion. Con esto el check 10 construye la imagen y el
  contenedor responde.
- El grep de generadores prohibidos del check 5 solo mira .cpp/.hpp y descarta
  lineas de comentario: la documentacion tiene que poder nombrar lo que
  prohibe. Es un estrechamiento del check y queda anotado en STATE.md.
- El pack implementar-regla se pasaba del presupuesto por tarea tras formatear
  rules.cpp: se saca state.hpp de la carga inicial en vez de subir el techo.

Gate: 11 checks PASS, solo falla el 9 (lint del loop), que necesita ledgers.

+++ engine/include/engine/rules.hpp
@@ -7,6 +7,7 @@
 /// Instanciado explicitamente para 7x7, 11x11 y 19x19 en `engine/src/rules.cpp`.
 
 #include <array>
+#include <cstddef>
 #include <span>
 
 #include <engine/state.hpp>
@@ -18,9 +19,8 @@ namespace engine {
 /// **rango promediado** (dos simultaneas en 3o/4o dan 3.5 a ambas): el motor oficial
 /// no les asigna orden, asi que desempatar por indice o por asiento seria inventarlo.
 /// ver docs/rules.md#r-12
-template <int MaxSnakes>
-struct PlacementsT {
-    std::array<float, MaxSnakes> rank{};
+template <int MaxSnakes> struct PlacementsT {
+    std::array<float, static_cast<std::size_t>(MaxSnakes)> rank{};
     int count{};
 };
 
@@ -69,8 +69,7 @@ template <int W, int H, int MaxSnakes>
 /// la semilla, asi que en partida real el lado del proximo shrink no es conocible.
 /// ver docs/rules.md#r-09
 template <int W, int H>
-[[nodiscard]] Bitboard<W, H> royale_hazards(std::uint64_t seed, int turn,
-                                            int shrink_every_n_turns);
+[[nodiscard]] Bitboard<W, H> royale_hazards(std::uint64_t seed, int turn, int shrink_every_n_turns);
 
 extern template Direction default_move<7, 7, 4>(const GameState<7, 7, 4>&, SnakeId) noexcept;
 extern template Direction default_move<11, 11, 4>(const GameState<11, 11, 4>&, SnakeId) noexcept;
